Update README, add new documentation files, refactor pom.xml
The README.md file was updated to refine the Java version recommendation. New documentation files have been added in both English and German explaining the Maven project file and the process of image creation, contributing to the project's transparency and usability. Additionally, properties were added to the pom.xml for PMD and Spotbugs plugins, and the option was implemented to determine whether the build should fail if issues are found. This improves the customizability and robustness of the build process.